Revert MainView Refactor / Verify valid selection
Reverted MainView refactor. fxml requires primary controller/fxml to be in Main package.

Add selection verification before dialogs will work.
